# public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Desafio Back API",
        "description": "Documentação da API do Desafio Back",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000/api/v1",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Autenticar Cliente",
                "description": "Faz login do cliente e retorna um token de autenticação",
                "operationId": "loginCliente",
                "requestBody": {
                    "description": "Dados de login",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Cliente"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login bem-sucedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "1|abcdefg123456..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro na autenticação"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/clientes": {
            "post": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Registrar novo cliente",
                "description": "Cria um novo cliente com email e senha",
                "operationId": "storeCliente",
                "requestBody": {
                    "description": "Dados do cliente",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Cliente"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cliente criado com sucesso"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "exemplo@email.com"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-04-20T14:30:00Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-04-20T14:30:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The email field is required. (and 1 more error)"
                                        },
                                        "errors": {
                                            "properties": {
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Invalid domain, email must be Gmail."
                                                    }
                                                },
                                                "password": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The password field must be at least 6 characters."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tools": {
            "get": {
                "tags": [
                    "Tools"
                ],
                "summary": "Lista ferramentas (com ou sem filtro por tags)",
                "operationId": "bd3375cee38e7b988f84e622337697b5",
                "parameters": [
                    {
                        "name": "tag",
                        "in": "query",
                        "description": "Filtrar ferramentas por tags (separadas por vírgula, exemplo: 'tag1,tag2')",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de ferramentas (filtradas ou não)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Tool"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tools"
                ],
                "summary": "Cria uma nova ferramenta",
                "operationId": "c827e9e594014e1ad7d329a872e0c35f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Tool"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ferramenta criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Tool"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao criar ferramenta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao criar ferramenta"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Detalhes do erro"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tools/{id}": {
            "delete": {
                "tags": [
                    "Tools"
                ],
                "summary": "Exclui uma ferramenta pelo ID",
                "operationId": "48cc66caf839068fd222bec8fcff3449",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da ferramenta a ser excluída",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ferramenta excluída com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ferramenta excluída com sucesso."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ferramenta não encontrada"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Cliente": {
                "title": "Cliente",
                "description": "Modelo de Cliente",
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "xxxxx@gmail.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "123Abc"
                    }
                },
                "type": "object"
            },
            "Tool": {
                "title": "Tool",
                "description": "Modelo de uma ferramenta",
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Nova ferramenta"
                    },
                    "link": {
                        "type": "string",
                        "example": "http://example.com"
                    },
                    "description": {
                        "type": "string",
                        "example": "Descrição da ferramenta"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "tag1",
                            "tag2"
                        ]
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Cliente",
            "description": "Cliente"
        },
        {
            "name": "Tools",
            "description": "Tools"
        }
    ]
}